organize types: Address
