+++ types/Address/index.d.ts
@@ -1,0 +1,5 @@
+export * from './Address';
+export * from './Verification';
+export * from './VerificationDetails';
+export * from './Verifications';
+export * from './AddressCreateParameters';
